SeedException report is displayed more often
SeedException report is now generated inside the toString() method instead of the
printStackTrace(), enabling it to be displayed in a lot more situations. Indeed,
most Web servers and runtime environment display exceptions with a toString().

In contrast, printStackTrace() will only print a classic stack trace without fancy
formatting.

When a SeedException is catched at the top of the application (in the servlet
listener, the CLI runner or the IT runner), it is no longer wrapped in an unexpected exception.
This improves clarity of exception reporting. If another type of exception is catched,
it is still wrapped.