Drop usage of ServletModule for container registration
Uses the ServletContainerInitializer to initialize and start the kernel which
allows to programmatically register filters, servlets and listeners directly
in plugins. As such the Web core module has been refactored to register all
those from other plugins with the WebProvider facet.

Servlet and filter registration through Guice servlet modules is no longer used.

Real servlet annotations (@WebServlet, @WebFilter and @WebListener) are now
supported with full injection. Global ordering of filters and listeners is possible
through the @Priority annotation or programmatically in their definitions.

As a side-effect this enables to have fully asynchronous filters chains and opens
the way to asynchronous REST resources.

Support for Servlet level 2.5 and below is dropped.