Java 9 compatibility, new testing API
Integration testing API is now based on SeedLauncher, allowing to leverage
UndertowLauncher, CliLauncher and potentially others to launch tests. This
results in testing that is much closer to a real application launch.

Arquillian testing is still supported but Undertow is now preferred for
Web integration tests. A new seed-web-tck module is provided to provide
a common test base for testing Web capabilities with various servers. This
is currently done with Tomcat through Arquillian and with Undertow.

Testing API has been refactored and improved, allowing to override configuration,
system properties, kernel parameters and arguments for each test (on the whole
test class or a particular test method).